analytics: rename from angulartics
